// clients/dealsService/models/DealUnitFlowCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DealUnitFlowCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    name?: string | null;
    description?: string | null;
    parentBusinessProcessId?: string | null;
    tenantId?: string | null;
    tenantEnrolmentId?: string | null;
};