// clients/dealsService/models/DealUnitFlowStageCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DealUnitFlowStageCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    order?: number;
    name?: string | null;
    dealUnitFlowId?: string | null;
    tenantId?: string | null;
    description?: string | null;
    enrolmentId?: string | null;
    parentBusinessProcessStageId?: string | null;
};