// clients/dealsService/models/DealUnitFlowStageUpdateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DealUnitFlowStageUpdateDto = {
    order?: number;
    name?: string | null;
    description?: string | null;
    enrolmentId?: string | null;
    dealUnitFlowId?: string | null;
    parentBusinessProcessStageId?: string | null;
};