// clients/dealsService/models/DealUnitFlowUpdateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DealUnitFlowUpdateDto = {
    name?: string | null;
    description?: string | null;
    parentBusinessProcessId?: string | null;
    tenantId?: string | null;
    tenantEnrollmentId?: string | null;
};